requirements.txt
snowflake-connector-python
requests
pyyaml
streamlit
altair
pandas
pillow
statistics
